refactor(mcp): deprecate file_export_mcp; switch to tools.server
- Replace tools/file_export_mcp.py with a deprecation stub pointing to
  tools/server.py (FastMCP) and utils/* modules
- Update MCPO config to run /app/.venv/bin/python -m tools.server; remove
  PYTHONPATH env
- Docker: include utils/ in image
- CI: build GHCR on folder-refac branch and add fc-latest tag

Why: consolidate MCP tooling under a new server architecture, modularize
utilities, improve container reproducibility, and enable CI for the refactor
branch.

BREAKING CHANGE: tools/file_export_mcp.py no longer exposes the previous API.
Migrate to tools/server.py and utils/* modules.
